models/predict/requirements.txt
opencv-python==4.4.0.44
opencv-contrib-python==4.4.0.44
numpy==1.19.5
tensorflow==2.4.1
flask==1.1.1
scikit-image==0.17.2
pillow==8.1.1
gunicorn==20.0.4
